feat: Widget waters only plant that needs water most. Don't overwater.
Water only the plant that closest to dying when clicking water drop
button in widget. If the plant doesn't need watering yet,
hide the drop button. Clicking the widget plant image takes you to
that plant's details screen. The widget image also displays the plant's id.

Still need to display the id of all plants on the main screen.