feat: Finish TWID.04-Exercise-Extras.
Change the water-all click event in the widget to water only that specific plant shown.
Hide the water drop button in the widget if it’s been less than 2 hours since it was last watered or the plant is dead.
Change the widget plant image click handler to launch the Detail Activity for that plant rather than the Main Activity.
Add a TextView underneath the plant image that displays the plant ID in the widget to help identify each plant.